### 2020-02-03 #368 @qdread
Presubmission inquiry: rslurm package
Submitting Author: Quentin Read (@qdread)
Repository:  https://github.com/sesync-ci/rslurm/

---

```
Package: rslurm
Title: Submit R Calculations to a 'Slurm' Cluster
Description: Functions that simplify submitting R scripts to a 'Slurm'
    workload manager, in part by automating the division of embarrassingly
    parallel calculations across cluster nodes.
Acknowledgements: Development of this R package was supported by the National
    Socio-Environmental Synthesis Center (SESYNC) under funding received from
    the National Science Foundation grants DBI-1052875 and DBI-1639145.
Version: 0.5.0
Authors@R: c(person('Philippe', 'Marchand', email = "marchand.philippe@gmail.com", role = 'aut'),
             person('Ian', 'Carroll', role = 'aut'),
             person('Mike', 'Smorul', role = 'ctb'),
             person('Rachael', 'Blake', role = 'ctb'),
             person('Quentin', 'Read', email = 'qread@sesync.org', role = c('ctb', 'cre')),
             person('Se Jong', 'Cho', role = 'art')
             )
    whisker (>= 0.3)
    parallel,

```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

The package allows users to write code that is executed in parallel on a [Slurm](https://slurm.schedmd.com/) cluster using `mapply` and `lapply`-like statements. Shell scripts are created and submitted behind the scenes so the user’s workflow can be completely contained in a single R script. It is essentially a wrapper for both the Slurm software and the parallel package. The rslurm package is classified as *scientific software wrapper* because its main use is to submit cluster jobs managed by Slurm software through R, and it is classified as *workflow automation* because it can be used to encapsulate an R and Slurm workflow into a single R script to facilitate automation.

The target audience is people at research institutions that have access to a high performance computing cluster with [Slurm](https://slurm.schedmd.com/) software installed. That now includes a lot of universities and government research institutions in multiple countries. Its scientific application is to simplify any type of research computing workflow done in R that includes embarrassingly parallel computations &mdash; instead of being split among multiple R scripts and shell scripts to set up the input data for the jobs (R), run the jobs (shell), then collect and process job output (back to R), it can all be done in a single R script.

Since our package was first released there is also another alternative that has been released, [slurmR](https://uscbiostats.github.io/slurmR/). I believe their functionality is largely overlapping and our package is at least as good as the other one in terms of both ease of use and performance. We are actively addressing issues and maintaining the package.

The package was originally written for people at our research center, [SESYNC](https://www.sesync.org), which is funded by the U.S. National Science Foundation. The original devs submitted it to CRAN a couple of years ago and it picked up a fair number of users there. It languished for a while without being updated but I recently took over active maintenance of the package. I have cleared out the queue of outstanding issues and added a couple of new functions to the package. I know that the user guide for ropensci suggests submitting here first before CRAN but obviously it’s already been on CRAN for a while and we think it would be appropriate for ropensci too.
