### 2020-02-09 #368 @maelle
Thanks @qdread for your pre-submission inquiry! The slurmR package has a [comparison table](https://uscbiostats.github.io/slurmR/index.html#vs) that seems to indicate it has more functionality than rslurm. Can you comment a bit on the differences? Thank you!
