### 2020-02-09 #368 @qdread
I think the functionality overlaps for the most part. rslurm now has functions corresponding to mapply() and lapply(), so I think it should also be listed as "yes" for the apply family (column 2 in the table). However they are correct that rslurm does not support restarting individual tasks that terminated with errors (column 1), nor does it support creating multinode clusters on Slurm so that functions from parallel can directly be used (column 3). We are considering adding the functionality to be able to rerun a subset of tasks from a job, but I don't think the multinode cluster feature is really necessary for our package. I think it's preferable to use job arrays.

In general I would argue that rslurm is simpler and more streamlined than slurmR, which I think makes it more user-friendly at the expense of possibly having a little fewer features. The goal in creating the package was to make using the Slurm cluster easier for users who are otherwise unfamiliar with how to parallelize their code.

So to sum up, I agree with their table but I think we have come down on the side of simplicity. The one feature I would consider adding would be support for rerunning jobs and automatically increasing memory or time limits for tasks that terminated due to running out of resources. That would be useful for debugging as well.
