### 2020-02-10 #368 @qdread
Following up on my previous answer here are two other points:

* rslurm has been on CRAN for several years and has a good proven record of usage in the community. It is downloaded a few hundred times a month and users are continually providing feedback and issues that have improved the package. I'm not sure whether slurmR has gotten tested or used by that many users.
* The comparison table on the slurmR page may not be fully accurate. I will open an issue with them to get them to update it to reflect the current state of rslurm.
