### 2020-02-24 #368 @qdread
Hi and sorry for pestering, but I was curious whether submitting the package is OK based on the response I gave to the request for more comment. Thanks!!!
