### 2020-02-24 #368 @maelle
Thanks for commenting and sorry for the misunderstanding: I'm actually waiting for your discussion with slurmR authors (to update their comparison table) before taking a new look.
