### 2020-02-24 #368 @qdread
The author of slurmR already updated the comparison table, which I think is now more or less correct.
