### 2020-02-26 c8fbb99
Edit error catching in slurm_apply and slurm_call They will not fail for functions with ellipsis in their formal arguments Also add some new contributors!
