+++ DESCRIPTION
@@ -17,6 +17,8 @@ Authors@R: c(person('Philippe', 'Marchand', email = "marchand.philippe@gmail.com
              person('Mike', 'Smorul', role = 'aut'),
              person('Rachael', 'Blake', role = 'aut'),
              person('Quentin', 'Read', email = 'qread@sesync.org', role = c('aut', 'cre')),
+             person('Dayne', 'Filer', role = 'ctb'),
+             person('Ben', 'Fasoli', role = 'ctb'),
              person('Se Jong', 'Cho', role = 'art')
              )
 Depends:
