### 2020-02-27 #368 @maelle
Thanks for providing more information.

 As per [our current policies](https://devguide.ropensci.org/policies.html#aims-and-scope), in particular [about overlap](https://devguide.ropensci.org/policies.html#overlap),

- more downloads isn't a criterion;

- broader functionality would be a criterion but slurmR has more functionality than rslurm;

- regarding usability and performance, we don't think one of the two packages is clearly better.

 Therefore we can't accept your package at this point. Thanks again for your pre-submission!
