### 2020-05-27 #368 @pmarchand1
Hello, I am sorry I didn't see this until now, I was one of the original authors of this package but have not been the maintainer for some time. To clarify, if there are two packages that perform one function, but neither is clearly the best, then neither could be on rOpenSci? This seems to be a worse outcome than even picking one of the two at random.
